# backend/api/openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  monitor.Monitor:
    properties:
      created_at:
        format: date-time
        type: string
      id:
        format: uuid
        type: string
      interval_seconds:
        type: integer
      updated_at:
        format: date-time
        type: string
      url:
        type: string
    required:
    - created_at
    - id
    - interval_seconds
    - updated_at
    - url
    type: object
  monitorcontroller.listResponse:
    properties:
      monitors:
        items:
          $ref: '#/definitions/monitor.Monitor'
        type: array
    required:
    - monitors
    type: object
  monitorcontroller.monitorRequest:
    properties:
      interval_seconds:
        example: 60
        maximum: 86400
        minimum: 60
        type: integer
      url:
        example: https://example.com
        maxLength: 2048
        type: string
    required:
    - interval_seconds
    - url
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    required:
    - code
    - message
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    required:
    - error
    type: object
info:
  contact: {}
  description: 'Uptime API. JSON bodies accept one object, reject unknown fields and
    are limited to 16 KiB. Requests with Origin require an allowed origin. Mutations
    require X-CSRF-Protection: 1.'
  title: Uptime API
  version: "1.0"
paths:
...
  /monitors:
    get:
      description: 'Accepts access_token cookie or Authorization: Bearer JWT. A present
        Authorization header takes precedence. Returns an empty array when there are
        no monitors.'
      operationId: monitorList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              description: no-store
              type: string
          schema:
            $ref: '#/definitions/monitorcontroller.listResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid credentials or token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Origin or CSRF rejected
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List monitors
      tags:
      - monitor
    post:
      consumes:
      - application/json
      description: 'Accepts access_token cookie or Authorization: Bearer JWT. A present
        Authorization header takes precedence. URL must use HTTP(S), without credentials
        or fragments.'
      operationId: monitorCreate
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/monitorcontroller.monitorRequest'
      - description: CSRF protection
        enum:
        - "1"
        in: header
        name: X-CSRF-Protection
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          headers:
            Cache-Control:
              description: no-store
              type: string
          schema:
            $ref: '#/definitions/monitor.Monitor'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid credentials or token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Origin or CSRF rejected
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Monitor URL already exists
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a monitor
      tags:
      - monitor
  /monitors/{id}:
    delete:
      description: 'Accepts access_token cookie or Authorization: Bearer JWT. A present
        Authorization header takes precedence.'
      operationId: monitorDelete
      parameters:
      - description: CSRF protection
        enum:
        - "1"
        in: header
        name: X-CSRF-Protection
        required: true
        type: string
      - description: Monitor UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          headers:
            Cache-Control:
              description: no-store
              type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid credentials or token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Origin or CSRF rejected
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Monitor not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a monitor
      tags:
      - monitor
    put:
      consumes:
      - application/json
      description: 'Accepts access_token cookie or Authorization: Bearer JWT. A present
        Authorization header takes precedence. URL must use HTTP(S), without credentials
        or fragments.'
      operationId: monitorUpdate
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/monitorcontroller.monitorRequest'
      - description: CSRF protection
        enum:
        - "1"
        in: header
        name: X-CSRF-Protection
        required: true
        type: string
      - description: Monitor UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              description: no-store
              type: string
          schema:
            $ref: '#/definitions/monitor.Monitor'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid credentials or token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Origin or CSRF rejected
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Monitor not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Monitor URL already exists
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a monitor
      tags:
      - monitor
securityDefinitions:
  BearerAuth:
    description: JWT with the Bearer prefix. Access-cookie is also accepted; a present
      Authorization header takes precedence even if invalid. Cookie authentication
      is described per operation because OpenAPI 2.0 has no cookie security scheme.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"